pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sm2can"
version = "0.2.0"
description = "Open-source macOS/Linux driver for SM2 Pro J2534 CAN adapters"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Aldo Guzman", email = "aldoguzman97@users.noreply.github.com"},
]
keywords = ["can-bus", "j2534", "sm2-pro", "scanmatik", "obd2", "automotive"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "pyusb>=1.2.0",
    "python-can>=4.0.0",
]

[project.optional-dependencies]
bluetooth = ["pybluez>=0.23"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.urls]
Homepage = "https://github.com/aldoguzman97/sm2can"
Documentation = "https://aldoguzman97.github.io/sm2can"
Repository = "https://github.com/aldoguzman97/sm2can"
Issues = "https://github.com/aldoguzman97/sm2can/issues"

[project.scripts]
sm2can = "sm2can.cli:main"
sm2can-probe = "sm2can.tools.probe:main"
sm2can-capture = "sm2can.tools.capture_decoder:main"

[project.entry-points."can.interface"]
sm2 = "sm2can.can_interface:SM2Bus"

[tool.setuptools.packages.find]
include = ["sm2can*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
